Fix tests after moving reporting out of k8s
